Item pickup measures from the body, not from the eye
The counters added in Phase 12 found this in their first session: broke 2,
placed 0, collected 0, two items still on the ground at exit. Pickup was a
1.4-block sphere measured from the camera, which holds the eye at 1.62; an item
rests at 0.12. Standing on an item put it 1.50 away, so the walk-over step of
the break-drop-collect-place loop had never worked -- since 7.10, which shipped
claiming that loop was closed.

The fix is the reference point rather than the radius. ItemEntities::PickupVolume
is a vertical segment from the feet to the top of the head; distanceSquaredTo
clamps onto it before measuring, the way vanilla's bounding box does. Enlarging
the radius would have hidden the symptom and left the shape of the error.

Most of the work was making the relationship testable. Six existing cases covered
ItemEntities well and none could have caught this, because each chose its own
reference point and its own radius -- the bug lived between Engine::kPickupRadius
and CharacterRenderer::kEyeHeight, two constants in two modules combined at one
call site. So kPickupRadius moves into ItemEntities where a test can reach it,
the point-taking overloads are deleted rather than kept, and Engine::playerFeet()
replaces the subtraction that was written out at four call sites before a fifth
caller passed the eye instead.

Three new cases pin it using the engine's real constants, one of them requiring
the eye-relative reach to still fail. 224 tests, asan and tsan clean.

Also fixes the asan preset, which had failed to *build* since the HUD landed:
HudRenderer's u16 digit bitmap promotes to int and tripped -Wsign-conversion.
Nothing noticed because the debug tree never recompiled that file.

Docs: DESIGN.md 7.14 has the write-up, and its phase plan drops three stale
claims -- that water is still upcoming, that Phase 9 is next, and the absent
Phase 15 that Engine.hpp has referenced since hearts landed.

Not verified in play. That is the next session's first job.

+++ src/render/HudRenderer.cpp
@@ -76,7 +76,10 @@ std::vector<u8> buildGlyphAtlas(u32 size, u32 count) {
             for (u32 column = 0; column < kDigitWidth; ++column) {
                 const u32 bit = (kDigitHeight - 1 - row) * kDigitWidth
                               + (kDigitWidth - 1 - column);
-                if ((kDigits[glyph] >> bit & 1u) == 0u) {
+                // Widened before the shift. `kDigits` is u16, so it promotes to a
+                // signed int and `& 1u` then converts back -- which `-Wsign-conversion`
+                // rejects, and rightly: the whole point of this bitmap is unsigned.
+                if (((static_cast<u32>(kDigits[glyph]) >> bit) & 1u) == 0u) {
                     continue;
                 }
                 blit(pixels, size, glyph, digitX + column, digitY + row);
